Am I casting at the wrong time and creating precision issues?

+++ aoc-10-02/src/main.rs
@@ -46,8 +46,8 @@ impl Grid {
         let locations: Vec<(usize, usize)> = self.asteroid_locations.keys().cloned().collect();
         for to in locations {
 
-            let x_diff = to.0 as f64 - self.laser_location.0 as f64;
-            let y_diff = to.1 as f64 - self.laser_location.1 as f64;
+            let x_diff = (to.0 as isize - self.laser_location.0 as isize) as f64;
+            let y_diff = (to.1 as isize - self.laser_location.1 as isize) as f64;
 
             let angle = (y_diff / x_diff).atan();
             self.asteroid_locations.insert(to, angle);
